Replace Identifier SyntaxKind with Atom
This commit also allows Atom to contain _any_ kind of text. However,
other groups will take priority from it, so this does not cause a
problem. I did this because command invocations are actually paths,
which are arbitrary text. Arguments to commands (i.e. bare words) will
also be Atoms.